👁🐀 Checkpoint ./komutlar/codeekle.js:4405668/185

+++ komutlar/codeekle.js
@@ -8,8 +8,10 @@ exports.run = (client, message, args) => {
   
    let kanal = args[0]
   let code = args.slice(1).join(' ');
+  let description = args.slice(2).join(' ');
   if (!kanal) return message.channel.send("Kanal İsmi Yazmalısın!");
   if (!code) return message.channel.send("Kodunu Yazmalısın!");
+  if (!description) return message.channel.send("Bir açıklama yazmayacanmı?")
   message.delete();
 if (message.guild.channels.find(a => a.name === "JavaScript")) {
 message.guild.createChannel(kanal, {type: "text", parent: message.guild.channels.find(a => a.name === "JavaScript")}).then(c => c.send(
@@ -23,6 +25,9 @@ ${client.emojis.get("669699664039116800")} İşte Kod;**
 
 ${code}
 
+---
+
+${description}
 
 **:heart: Sayın Kullanıcılar Kodda Herhangi Bir Hata Varsa <#669656253806673939> Kanalı Yardımıyla Yetkililere İletin!\n:heart: İyi Günler!**`))
